+++ CHANGELOG.md
@@ -5,6 +5,7 @@ next
 
 - Rename `queries` module into `statements`, and use that terminology everywhere
 - Add a new convenience function called `from` to create `SELECT` statements by starting from the `FROM` clause
+- Fix a broken link in `WithClause` documentation
 
 0.4.0
 -----